Add clang format
