+++ .clang-format
@@ -1,0 +1,3 @@
+Language:        Cpp
+BasedOnStyle:  Google
+PointerAlignment: Left
