+++ src/main_test.cpp
@@ -1,6 +1,6 @@
 #include "../third_party/googletest/googletest/include/gtest/gtest.h"
 
 int main(int argc, char **argv) {
-    ::testing::InitGoogleTest(&argc, argv);
-    return RUN_ALL_TESTS();
+  ::testing::InitGoogleTest(&argc, argv);
+  return RUN_ALL_TESTS();
 }
